Try to add MPD support
