+++ .github/workflows/deploy-docs.yml
@@ -35,17 +35,17 @@ jobs:
           pip install mkdocs mkdocs-material mkdocs-ivory
 
       - name: Prepare documentation directory
-        # This crucial step copies the root README.md to web/index.md.
+        # This crucial step copies the root README_full.md to web/index.md.
         # MkDocs, by default, looks for 'index.md' in its configured 'docs_dir'
         # (which is likely 'web' in your mkdocs.yml).
-        # This makes your root README.md the homepage of your MkDocs site.
+        # This makes your root README_full.md the homepage of your MkDocs site.
         run: |
           mkdir -p web # Ensure the 'web' directory exists before copying
-          cp README.md web/index.md # Copy the root README.md and rename it to index.md inside 'web'
+          cp README_full.md web/index.md # Copy the root README_full.md and rename it to index.md inside 'web'
 
       - name: Build MkDocs documentation
         # This command builds your site into the './site' directory by default.
-        # It will now use 'web/index.md' (which is your copied README.md) as the homepage.
+        # It will now use 'web/index.md' (which is your copied README_full.md) as the homepage.
         # Ensure your mkdocs.yml file has 'docs_dir: web' if your source markdown files are in 'web'.
         run: mkdocs build
