+++ config.py
@@ -2,7 +2,7 @@
 import os
 
 # --- Media Server Type ---
-MEDIASERVER_TYPE = os.environ.get("MEDIASERVER_TYPE", "jellyfin").lower() # Possible values: jellyfin, navidrome, lyrion
+MEDIASERVER_TYPE = os.environ.get("MEDIASERVER_TYPE", "jellyfin").lower() # Possible values: jellyfin, navidrome, lyrion, mpd
 
 # --- Jellyfin and DB Constants (Read from Environment Variables first) ---
 
@@ -23,9 +23,16 @@ NAVIDROME_PASSWORD = os.environ.get("NAVIDROME_PASSWORD", "your_navidrome_passwo
 # These are used only if MEDIASERVER_TYPE is "lyrion".
 LYRION_URL = os.environ.get("LYRION_URL", "http://your_lyrion_url:9000")
 
+# --- MPD (Music Player Daemon) Constants ---
+# These are used only if MEDIASERVER_TYPE is "mpd".
+MPD_HOST = os.environ.get("MPD_HOST", "localhost")
+MPD_PORT = int(os.environ.get("MPD_PORT", "6600"))
+MPD_PASSWORD = os.environ.get("MPD_PASSWORD", "")  # Optional password, leave empty if none
+MPD_MUSIC_DIRECTORY = os.environ.get("MPD_MUSIC_DIRECTORY", "/var/lib/mpd/music")  # Path to MPD's music directory for file access
+
 
 # --- General Constants (Read from Environment Variables where applicable) ---
-APP_VERSION = "v0.6.8-beta" # Version updated for Sonic Fingerprint
+APP_VERSION = "v0.6.10-beta" # Version updated for Sonic Fingerprint
 MAX_DISTANCE = 0.5
 MAX_SONGS_PER_CLUSTER = 0
 MAX_SONGS_PER_ARTIST = int(os.getenv("MAX_SONGS_PER_ARTIST", "3")) # Max songs per artist in similarity results and clustering
